% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bombay_1951.R
\docType{data}
\name{bombay_1951}
\alias{bombay_1951}
\title{Bombay population totals in 1941 and 1951 by age}
\format{
Data frame with 13 rows and 5 columns:
\describe{
  \item{age_1941}{Age group in 1941}
  \item{age_1951}{Age group in 1951}
  \item{pop_1941}{Enumerated population in 1941}
  \item{pop_1951}{Enumerated population in 1951}
  \item{sr}{Census survival ratio derived from the United Nations model life table corresponding to a life expectancy at birth of45 years for males. See Manual III: Methods for Popu- lation Projections by Sex and Age (United Nations publication, Sales No.: 56.XIII.3).}
}
}
\source{
Indian Population Census. Published in United Nations Department of Economic and Social Affairs Population Division - 1983 - Methods of measuring internal migration \url{https://www.un.org/en/development/desa/population/publications/manual/migration/measuring-migration.asp}
}
\usage{
bombay_1951
}
\description{
Population data for Bombay by age in 1941 and 1951
}
\keyword{datasets}
