% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffs_rates.R
\name{ffs_rates}
\alias{ffs_rates}
\title{Estimation of Bilateral Migrant Flows from Bilateral Migrant Stocks Using Rates Approaches}
\usage{
ffs_rates(m1 = NULL, m2 = NULL, M = NULL, method = NULL)
}
\arguments{
\item{m1}{Matrix of migrant stock totals at time \emph{t}. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t}}

\item{m2}{Matrix of migrant stock totals at time \emph{t}+1. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t}+1.}

\item{M}{Numeric value for the global sum of migration flows, used for \code{dennett} approach.}

\item{method}{Method to estimate flows. Can take values \code{dennett} or \code{rogers-von-rabenau}. See detials section for more information.}
}
\value{
Estimates migrant transitions flows based on migration rates.

When \code{method = "dennett"} migration are derived from the matrix supplied to \code{m1}. Dennett uses bilateral migrant stocks at begining of period. Rates then multiplied by global migration flows supplied in \code{M}.

When \code{method = "rogers-von-rabenau"} a matrix of growth rates are derived from the changes in inital poplations stock \code{m1} to obtain \code{m2};
\deqn{P^{t+1} = g P^{t}}
and then multiplied by the corresponding populations at risk in \code{m1}. Can result in negative flows.
}
\description{
Estimates migrant transitions flows between two sequential migrant stock tables using approached based on rates.
}
\examples{
P1 <- matrix(data = c(1000, 100, 10, 0, 55, 555, 50, 5, 80, 40, 800, 40, 20, 25, 20, 200),
             nrow = 4, ncol = 4, byrow = TRUE)
P2 <- matrix(data = c(950, 100, 60, 0, 80, 505, 75, 5, 90, 30, 800, 40, 40, 45, 0, 180),
             nrow = 4, ncol = 4, byrow = TRUE)
reg <- LETTERS[1:4]
dimnames(P1) <- dimnames(P2) <- list(pob = reg, por = reg)
P1; P2

# calculate totol migration flows for dennett approach
n <- colSums(P2) - colSums(P1)

ffs_rates(m1 = P1, M =  sum(abs(n)), method = "dennett" )
ffs_rates(m1 = P1, m2 = P2, method = "rogers-von-rabenau" )
}
\seealso{
\code{\link{ffs_demo}}, \code{\link{ffs_rates}}
}
\author{
Guy J. Abel
}
