% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_migration_tibble.R
\name{format_migration_tibble}
\alias{format_migration_tibble}
\title{Helper function to format migration input}
\usage{
format_migration_tibble(
  m,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow"
)
}
\arguments{
\item{m}{A \code{matrix} or data frame of origin-destination flows. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to \code{orig_col}, \code{dest_col} and \code{flow_col}.}

\item{orig_col}{Character string of the origin column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{dest_col}{Character string of the destination column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{flow_col}{Character string of the flow column name (when \code{m} is a data frame rather than a \code{matrix})}
}
\value{
Formatted tibble
}
\description{
Helper function to format migration input
}
