% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_distance.R
\name{index_distance}
\alias{index_distance}
\title{Summary indices of migration distance}
\source{
Bell, M., Blake, M., Boyle, P., Duke-Williams, O., Rees, P. H., Stillwell, J., & Hugo, G. J. (2002). Cross-national comparison of internal migration: issues and measures. Journal of the Royal Statistical Society: Series A (Statistics in Society), 165(3), 435–464. https://doi.org/10.1111/1467-985X.00247
}
\usage{
index_distance(
  m = NULL,
  d = NULL,
  orig = "orig",
  dest = "dest",
  flow = "flow",
  dist = "dist",
  long = TRUE
)
}
\arguments{
\item{m}{A \code{matrix} or data frame of origin-destination flows. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to \code{orig}, \code{dest} and \code{flow}.}

\item{d}{A \code{matrix} or data frame of origin-destination distances. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to \code{orig}, \code{dest} and \code{dist}. Region names should match those in \code{m}.}

\item{orig}{Character string of the origin column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{dest}{Character string of the destination column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{flow}{Character string of the flow column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{dist}{Character string of the distance column name (when \code{dist} is a data frame rather than a \code{matrix})}

\item{long}{Logical to return a long data frame with index values all in one column}
}
\value{
A tibble with 3 summary measures where
\item{mean}{Mean migration distance from Bell et. al. (2002) - not discussed in text but given in Table 6}
\item{median}{Mean migration distance from Bell et. al. (2002)}
\item{decay}{Distance decay parameter obtained from a Poisson regression model (\code{flow ~ orig + dest + log(dist)})}
}
\description{
Summary indices of migration distance
}
\examples{
# single year
index_distance(
  m = subset(korea_gravity, year == 2020),
  d = subset(korea_gravity, year == 2020),
  dist = "dist_cent"
)

# multiple years
library(dplyr)
library(tidyr)
library(purrr)

korea_gravity \%>\%
  select(year, orig, dest, flow, dist_cent) \%>\%
  group_nest(year) \%>\%
  mutate(i = map2(
    .x = data, .y = data,
    .f = ~index_distance(m = .x, d = .y, dist = "dist_cent", long = FALSE)
  )) \%>\%
  select(-data) \%>\%
  unnest(i)
}
