% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indian_sub.R
\docType{data}
\name{indian_sub}
\alias{indian_sub}
\title{Lifetime migration totals for states and zones in the Indian 1901 to 1931}
\format{
Data frame with 164 rows and 7 columns:
\describe{
  \item{zone}{Zone of state. In somce cases the state and zone are the same entity}
  \item{state}{Indian state}
  \item{sex}{Migrant sex}
  \item{in_migrants}{In-migrant total based on birthplace}
  \item{out_migrants}{Out-migrant total based on birthplace}
  \item{net_migrants}{Net migrant total based on birthplace}
}
}
\source{
A Historical Study Of Internal Migration In The Indian Sub-continent 1901-1931 by Zachariah, K.C. 
Scaped from \url{https://archive.org/details/in.ernet.dli.2015.130424/page/n73/mode/2up}
}
\usage{
indian_sub
}
\description{
Lifetime migration (stock) totals from India
}
\keyword{datasets}
