% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korea_dist.R
\docType{data}
\name{korea_dist}
\alias{korea_dist}
\title{Distances in kilometers between centroids of Korean administrative regions}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 17 rows and 17 columns.
}
\usage{
korea_dist
}
\description{
Distance matrix of kilometers between centroids of first level administrative regions of South Korea.
}
\details{
Data derived in R using the GADM shape files (downloaded using raster::getData), sf::st_centroid to calculate region centroids, and geosphere::distm to calculate great circle distance between centroid points
}
\keyword{datasets}
