% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korea_pop.R
\docType{data}
\name{korea_pop}
\alias{korea_pop}
\title{Annual resident population totals of Korean regions}
\format{
Data frame with 2,601 rows and 4 columns:
\describe{
  \item{region}{Administrative region}
  \item{year}{Year of flow}
  \item{population}{Resident Population}
}
}
\source{
Source: Statistics Korea, Population Statistics Based on Resident Registration. Data downloaded from \url{https://kosis.kr/eng} in July 2021.
}
\usage{
korea_pop
}
\description{
Annual resident population totals between 2012 and 2020 based on first level administrative regions.
}
\keyword{datasets}
