% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korea_reg.R
\docType{data}
\name{korea_reg}
\alias{korea_reg}
\title{Annual origin destination migration flows between Korean regions}
\format{
Data frame with 2,601 rows and 4 columns:
\describe{
  \item{orig}{Origin region}
  \item{dest}{Destination region}
  \item{year}{Year of flow}
  \item{flow}{Migration flow}
}
}
\source{
Statistics Korea, Internal Migration Statistics. Data downloaded from \url{https://kosis.kr/eng} in July 2021.
}
\usage{
korea_reg
}
\description{
Origin-destination migration flows between 2012 and 2020 based on first level administrative regions.
}
\keyword{datasets}
