% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rc_un.R
\docType{data}
\name{model_rc_un}
\alias{model_rc_un}
\title{Model parameters for six Rogers-Castro migration schedules proposed by UN DESA}
\format{
A \code{tibble} with five columns and 84 rows:
\describe{
  \item{schedule}{Character string for full name of schedule}
  \item{value}{Character string for abbrivated name of schedule}
  \item{param}{Character string for sex of schedule}
  \item{param}{Character string for the seven parameters}
  \item{value}{Parameter values}
}
}
\source{
United Nations Department of Economic and Social Affairs Population Division. (1992). Preparing Migration Data for Subnational Population Projections. http://www.un.org/esa/population/techcoop/IntMig/migdata_popproj/migdata_popproj.html
}
\usage{
model_rc_un
}
\description{
Sets of parameters for the Rogers-Castro migration age schedule proposed by UN DESA
}
\keyword{datasets}
