% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchars_wrap.R
\name{nchars_wrap}
\alias{nchars_wrap}
\title{Count the number of characters per line}
\usage{
nchars_wrap(b, w)
}
\arguments{
\item{b}{Numeric vector for the position of line breaks between the words in \code{w}}

\item{w}{Character string vector of words}
}
\value{
List with vectors for number of characters per line and the number of words per line
}
\description{
Count the number of characters per line
}
