% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_param.R
\name{net_param}
\alias{net_param}
\title{Estimate Parameters for Net Migration Scaling.}
\usage{
net_param(m, region, ntot)
}
\arguments{
\item{m}{Matrix of origin-destination flows, where the first and second dimensions correspond to origin and destination respectively.}

\item{region}{Integer value corresponding to the region that the net migration sum is desired. Will return sums for all regions by default.}

\item{ntot}{Vector of net migraiton totals to constrain the sum of the imputed cell columns. Elements must sum to zero.}
}
\value{
Vector of two values corresonding to the roots fo the quatratic equation.
}
\description{
This function is predominantly intended to be used within the \code{\link{ipf_net}} routine.
}
\author{
Guy J. Abel
}
