% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_scale.R
\name{net_scale}
\alias{net_scale}
\title{Scale Migration Flows in Origin-Destination}
\usage{
net_scale(m, region = NULL, alpha)
}
\arguments{
\item{m}{Matrix of origin-destination flows, where the first and second dimensions correspond to origin and destination respectively.}

\item{region}{Integer corresponding to row (column) in a square matrix for the region where scaling is to be applied}

\item{alpha}{Numeric value of the scaling factor}
}
\value{
Matrix scaled in region(s) by value of alpha, where \code{alpha} applied to destination flows and inverse of \code{alpha} applied to origin flows
}
\description{
This function is predominantly intended to be used within the \code{\link{cm_net}} routine.
}
\author{
Guy J. Abel
}
