% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_stream.R
\name{net_stream}
\alias{net_stream}
\title{Calculate Counterstream and Stream Net Migration}
\usage{
net_stream(m)
}
\arguments{
\item{m}{A \code{matrix} or data frame of origin-destination flows. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame the first and second columns correspond to origin and destination respecivly. The third column contains the number of migraitons or migrants.}
}
\value{
A \code{tibble} with columns for orig, destination, corridor, stream, countersteam and net stream in each pair (corridor)
}
\description{
Calculate Counterstream and Stream Net Migration
}
