% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_england_1960.R
\docType{data}
\name{new_england_1960}
\alias{new_england_1960}
\title{New England male white-native population totals in 1950 and 1960 by place of birth and age}
\format{
Data frame with 72 rows and 4 columns:
\describe{
  \item{birthplace}{Place of birth (US Census area)}
  \item{year}{Year}
  \item{age_1960}{Age group in 1960}
  \item{pop_1950}{Enumerated population in 1950}
  \item{pop_1960}{Enumerated population in 1960}
}
}
\source{
United States Bureau of the Census, United States Census of Population: 1960..Subject Reports.."State of birth" (Washington, D.C.), table 25, pp. 61-62. Persons with place of birth not reported were distributed pro rata among those with place of birth reported. 

Published in United Nations Department of Economic and Social Affairs Population Division. (1970). Methods of measuring internal migration. United Nations Department of Economic and Social Affairs Population Division - 1970 - Methods of measuring internal migration \url{https://www.un.org/development/desa/pd/sites/www.un.org.development.desa.pd/files/files/documents/2020/Jan/manual_vi_methods_of_measuring_internal_migration.pdf}
}
\usage{
new_england_1960
}
\description{
New England population data for by place of birth and age in 1950 and 1960 for male white native born.
}
\keyword{datasets}
