% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_model_fund.R
\docType{data}
\name{rc_model_fund}
\alias{rc_model_fund}
\title{Fundamental parameters for Rogers-Castro migration schedule}
\format{
A \code{tibble} with two columns and seven rows:
\describe{
  \item{param}{Character string for the seven parameters}
  \item{value}{Parameter values}
}
}
\source{
Rogers, A., and L. J. Castro. (1981). Model Migration Schedules. \emph{IIASA Research Report 81} RR-81-30
}
\usage{
rc_model_fund
}
\description{
Set of fundamental parameters for the Rogers-Castro migration age schedule, as suggested in Rogers and Castro (1981).
}
\keyword{datasets}
