% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_wrap_n_single.R
\name{str_wrap_n_single}
\alias{str_wrap_n_single}
\title{Single line wrap for string}
\usage{
str_wrap_n_single(string = NULL, n = 2)
}
\arguments{
\item{string}{string from \code{str_wrap_n}}

\item{n}{n from from \code{str_wrap_n}}
}
\value{
String with line breaks
}
\description{
Single line wrap for string
}
