% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_bilateral.R
\name{sum_bilateral}
\alias{sum_bilateral}
\title{Summary of bilateral flows, counter-flow and net migration flow}
\usage{
sum_bilateral(
  m,
  label = "flow",
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow"
)
}
\arguments{
\item{m}{A \code{matrix} or data frame of origin-destination flows. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to \code{orig_col}, \code{dest_col} and \code{flow_col}.}

\item{label}{Character string for the prefix of the calculated columns. Can take values \code{flow} or \code{stream}}

\item{orig_col}{Character string of the origin column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{dest_col}{Character string of the destination column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{flow_col}{Character string of the flow column name (when \code{m} is a data frame rather than a \code{matrix})}
}
\value{
A \code{tibble} with columns for orig, destination, corridor, flow, counter-flow and net flow in each bilateral pair.
}
\description{
Summary of bilateral flows, counter-flow and net migration flow
}
\examples{
# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
sum_bilateral(m)

# data frame
library(dplyr)
library(tidyr)
d <- expand_grid(orig = r, dest = r, sex = c("female", "male")) \%>\%
  mutate(flow = sample(x = 1:100, size = 32))
d

# use group_by to distinguish od tables
d \%>\%
  group_by(sex) \%>\%
  sum_bilateral()
}
