% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_expand.R
\name{sum_expand}
\alias{sum_expand}
\title{Expand flow data to include aggregate sums for origin and destination meta-regions}
\usage{
sum_expand(
  m,
  return_matrix = TRUE,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  orig_area_col = "orig_area",
  dest_area_col = "dest_area",
  orig_area = NULL,
  dest_area = NULL
)
}
\arguments{
\item{m}{A \code{matrix} or data frame of origin-destination flows. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to \code{orig_col}, \code{dest_col} and \code{flow_col}.}

\item{return_matrix}{Logical to return a matrix. Default `TRUE`}

\item{orig_col}{Character string of the origin column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{dest_col}{Character string of the destination column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{flow_col}{Character string of the flow column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{orig_area_col}{Character string of the origin area column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{dest_area_col}{Character string of the destination area column name (when \code{m} is a data frame rather than a \code{matrix})}

\item{orig_area}{Vector of labels for the origin areas of each row of \code{m}.}

\item{dest_area}{Vector of labels for the destination areas of each row of \code{m}.}
}
\value{
A \code{tibble} or \code{matrix} with additional row and columns (for matrices) for aggregate sums for origin and destination meta-regions
}
\description{
Expand matrix of data frame of migration data to include aggregate sums for corresponding origin and destination meta regions.
}
\examples{
##
## from data frame
##
library(tidyverse)
d <- block_matrix(x = 1:25, b = c(2,3,4,2,1)) \%>\%
  as.data.frame.table(responseName = "flow") \%>\%
  as_tibble() \%>\%
  rename(orig = 1, 
         dest = 2) \%>\%
  mutate(orig_area = str_sub(string = orig, end = 1),
         dest_area = str_sub(string = dest, end = 1))
d

sum_expand(d)

# return a matrix
sum_expand(d, return_matrix = FALSE)

##
## from matrix
##
m <- block_matrix(x = 1:16, b = c(2,3,4,2))
m
# requires a vector of origin and destination areas
a <- rep(LETTERS[1:4], times = c(2,3,4,2))
a
sum_expand(m = m, orig_area = a, dest_area = a)
}
