% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_total.R
\name{sum_tidy_total}
\alias{sum_tidy_total}
\title{Origin or Destination Sums for Origin-Destination Tidy Data}
\usage{
sum_tidy_total(d = NULL, type = "orig")
}
\arguments{
\item{d}{data frame of origin-destination data. Must contain columns with names orig, dest and flow.}

\item{type}{string (either orig or dest) indicating if totals should be for origins (outflows) or destinations (inflows)}
}
\value{
\code{tibble} data frame
}
\description{
Sums of inflows and out flows from tidy origin-destination data.
}
\author{
Guy J. Abel
}
