% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrella.R
\docType{data}
\name{umbrella}
\alias{umbrella}
\title{Umbrella colour scheme}
\format{
An object of class \code{character} of length 9.
}
\usage{
umbrella
}
\description{
Vector of hexadecimal codes for a umbrella rainbow colour scheme
}
\keyword{datasets}
