% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffs_diff.R
\name{ffs_diff}
\alias{ffs_diff}
\title{Estimation of Bilateral Migrant Flows from Bilateral Migrant Stocks Using Stock Differencing Approaches}
\usage{
ffs_diff(m1, m2, decrease = "return")
}
\arguments{
\item{m1}{Matrix of migrant stock totals at time \emph{t}. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t}}

\item{m2}{Matrix of migrant stock totals at time \emph{t}+1. Rows in the matrix correspond to place of birth and columns to place of residence at time \emph{t}+1.}

\item{decrease}{How to treat decreases in bilateral stocks over the \emph{t} to \emph{t}+1 period (so as to avoid a negative bilateral flow estimates). See details for possible options.}
}
\value{
Estimates migrant transitions flows between two sequential migrant stock tables. 

When \code{decrease = "zero"} all decreases in migrant stocks over there period are set to zero, following the approach of Bertoli and Fernandez-Huertas Moraga (2015) 

When \code{decrease = "return"} all decreases in migrant stocks are assumed to correspond to return flows back to their place of birth, following the approach of Beine and Parsons (2015)
#' @references 
Beine, Michel, Simone Bertoli, and Jesús Fernández-Huertas Moraga. (2016). A Practitioners’ Guide to Gravity Models of International Migration. \emph{The World Economy} 39(4):496–512.
}
\description{
Estimates migrant transitions flows between two sequential migrant stock tables using differencing approaches commonly used by economists.
}
\examples{
P1 <- matrix(data = c(1000, 100, 10, 0, 55, 555, 50, 5, 80, 40, 800, 40, 20, 25, 20, 200),
             nrow = 4, ncol = 4, byrow = TRUE)
P2 <- matrix(data = c(950, 100, 60, 0, 80, 505, 75, 5, 90, 30, 800, 40, 40, 45, 0, 180),
             nrow = 4, ncol = 4, byrow = TRUE)
reg <- LETTERS[1:4]
dimnames(P1) <- dimnames(P2) <- list(pob = reg, por = reg)
P1; P2

ffs_diff(m1 = P1, m2 = P2, decrease = "zero")
ffs_diff(m1 = P1, m2 = P2, decrease = "return")
}
\seealso{
\code{\link{ffs_demo}}, \code{\link{ffs_rates}}
}
\author{
Guy J. Abel
}
