% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/korea_dist.R
\docType{data}
\name{korea_dist}
\alias{korea_dist}
\title{Weighted distances in kilometers between 2020 population centroids of Korean administrative regions}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 17 rows and 17 columns.
}
\source{
Robin Edwards, Maksym Bondarenko, Andrew J. Tatem and Alessandro Sorichetta. Unconstrained subnational Population Weighted Density in 2000, 2005, 2010, 2015 and 2020 ( 100m resolution ). WorldPop, University of Southampton, UK.
}
\usage{
korea_dist
}
\description{
Distance matrix of kilometers between population weighted centroids in 2020 of first level administrative regions of South Korea.
}
\keyword{datasets}
