% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_eqn.R
\name{quadratic_eqn}
\alias{quadratic_eqn}
\title{Solve Quadratic Equation}
\source{
Adapted from https://rpubs.com/kikihatzistavrou/80124
}
\usage{
quadratic_eqn(a, b, c)
}
\arguments{
\item{a}{Numeric value for quadratic term of x.}

\item{b}{Numeric value for multiplicative term of x.}

\item{c}{Numeric value for constant term.}
}
\value{
Vector of two values corresonding to the roots fo the quatratic equation.
}
\description{
General function to solve classic quadratic equation:
\deqn{ a x^2 + b x + c = 0 }
}
\examples{
quadratic_eqn(a = 2, b = 4, c = -6)
}
\author{
Guy J. Abel
}
