% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_turnover.R
\name{sum_turnover}
\alias{sum_turnover}
\title{Extract Total In- and Out-Migration Totals from an Origin-Destination Migration Flow Matrix.}
\usage{
sum_turnover(m, drop_diagonal = TRUE, include_net = TRUE)
}
\arguments{
\item{m}{A \code{matrix} or data frame of origin-destination flows. For \code{matrix} the first and second dimensions correspond to origin and destination respectively. For a data frame the first and second columns correspond to origin and destination respecivly. The third column contains the number of migraitons or migrants.}

\item{drop_diagonal}{Logical to indicate dropping of diagonal terms, where the origin and destination are the same, in the calcuation of totals. Default \code{TRUE}.}

\item{include_net}{Logical to indicate inclusion of a net migration total column for each region, in addition to the total in- and out-flows. Default \code{TRUE}.}
}
\value{
A \code{tibble} with total in- and out-flows for each region.
}
\description{
Extract Total In- and Out-Migration Totals from an Origin-Destination Migration Flow Matrix.
}
