% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uar_1960.R
\docType{data}
\name{uar_1960}
\alias{uar_1960}
\title{Lifetime migration data for Governorates of United Arab Republic in 1960}
\format{
Matrix with 11 rows and columns
\describe{
  \item{orig}{Governorate of birth}
  \item{carat}{Governorate of enumeration}
}
}
\source{
United Arab Republic, Department of Statistics and Census, 1960 Census of Population (Cairo, July 1963), vol. II, General tables, table 14, p. 50. 

Published in United Nations Department of Economic and Social Affairs Population Division. (1970). Methods of measuring internal migration. United Nations Department of Economic and Social Affairs Population Division - 1970 - Methods of measuring internal migration \url{https://www.un.org/development/desa/pd/sites/www.un.org.development.desa.pd/files/files/documents/2020/Jan/manual_vi_methods_of_measuring_internal_migration.pdf}
}
\usage{
uar_1960
}
\description{
Lifetime migration (stock) bilateral data from Governorates of the United Arab Republic
}
\keyword{datasets}
